refactor(workspace): use public plugin authoring APIs
Move ask-user front/server imports onto public workspace barrels and boring-ui-kit primitives so the plugin boundary is closer to external-package ready.
