+++ packages/workspace/tsconfig.front.json
@@ -15,7 +15,8 @@
     "noEmit": true,
     "types": ["vitest/globals", "@testing-library/jest-dom"],
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@hachej/boring-workspace": ["./src/index.ts"]
     }
   },
   "include": ["src/**/*"],
