+++ packages/workspace/vite.config.ts
@@ -38,6 +38,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": resolve(__dirname, "src"),
+      "@hachej/boring-workspace": resolve(__dirname, "src/index.ts"),
     },
   },
   build: {
