+++ packages/workspace/vitest.config.ts
@@ -5,6 +5,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": resolve(__dirname, "src"),
+      "@hachej/boring-workspace": resolve(__dirname, "src/index.ts"),
     },
   },
   test: {
